Utils: Rework aspect implementation
This avoids some repetition and could be a step towards
having type storage in (or rather accessible from) the base,
so we can have aspects for more complex data (treemodels...)
that are not easily converted to QVariant.

+++ src/plugins/beautifier/clangformat/clangformatsettings.cpp
@@ -246,7 +246,7 @@ public:
                                     configurations, predefinedStyleButton] {
             const bool predefSelected = styleButtonGroup->checkedButton() == predefinedStyleButton;
             predefinedBlob->setEnabled(predefSelected);
-            fallbackBlob->setEnabled(predefSelected && s.predefinedStyle.volatileValue().toInt() == 5); // File
+            fallbackBlob->setEnabled(predefSelected && s.predefinedStyle.volatileValue() == 5); // File
             configurations->setEnabled(!predefSelected);
         };
         updateEnabled();
